Automatically validate the engines.node range using ls-engines
Add ls-engines[1] as a dependency and use it as part of vetting
(specifically the `vet:package.json` subcommand) to ensure that the
engines.node range specified for this project is in line with all of its
dependencies. That is, if the manifest claims the package is compatible
with a Node.js version not supported by one of the (production or peer)
dependencies, it will complain.

--
1. https://www.npmjs.com/package/ls-engines