Automatically validate the `engines.node` range using ls-engines (#1695)
* Automatically validate the engines.node range using ls-engines

Add ls-engines[1] as a dependency and use it as part of vetting
(specifically the `vet:package.json` subcommand) to ensure that the
engines.node range specified for this project is in line with all of its
dependencies. That is, if the manifest claims the package is compatible
with a Node.js version not supported by one of the (production or peer)
dependencies, it will complain.

--
1. https://www.npmjs.com/package/ls-engines

* Add exception for `jsonp` to licensee configuration

The package `jsonp` is licensed under the MIT license but does not
declare this in the project manifest. To verify the license run the
command:

    cat node_modules/jsonp/Readme.md

and look for the `## License` section.