Switch `eslint-plugin-jsonc` to `(at)eslint/json`
Following [1] this replaces `eslint-plugin-jsonc` by `(at)eslint/json`
for linting JSON(C) in this project. The motivation for this change is
supported by the FAQ of the new package.

The reduction in rules is further motivated by ESLint's general move
away from formatting-like rules to lint-like rules, which I believe
makes sense for the present project given the use of Prettier.

--
1. https://eslint.org/blog/2024/10/eslint-json-markdown-support/