Use `depreman` to manage deprecations (#1762)
It's an early prototype by me to attempt to manage deprecations. Using
it here to dogfeed it on my own projects and understand what needs to
change or be supported as well as just to start managing the deprecated
dependencies of this project.

This also shuffles around some scripts and CI jobs with the goal of
improved consistency. I renamed `license-check` to `audit:licenses` in
part because I think it sounds a bit better, but primarily because it is
listed under "Auditing" in the Contributing Guidelines. However, in
contrast to deprecations and vulnerabilities, the license check should
always pass so maybe it's better to group it under "Vetting"?

The license check for `depreman` is explicitly allowlist because it is
distributed under the AGPL license. This is fine for this devDependency,
but would not be okay for "runtime" dependencies, hence the exception.