Use matrix for `npm run check:*` commands in CI
Reduce duplication and the overall size of the CI configuration by using
the fact that all `check:*` commands are more or less identical and can
thus perfectly be used in a matricized CI job.

This also folds the lockfile and manifest checks (`package.json` and
`package-lock.json` in CI) into the main `checks.yml` workflow to
further simplify things (the previous setup was chosen to reduce wasting
CI cycles on checking these when the files didn't change, but the
simplicity of the approach here seems better to me atm).

Fingers crossed this isn't going to be a nightmare with
`step-security/harden-runner`.