Enable generic ESLint rules on whole code base consistently
Refactor the ESLint configuration primarily with the goal of enforcing a
consistent style on all JavaScript in the code base to the extend
possible.

This change did uncover some new problems with the newly introduced
unicorn plugin:

- `consistent-destructuring`: like `prefer-ternary`: too situational
  to lint for in my opinion. Open to reconsider later.
- `number-literal-case`: conflicts with Prettier

Lastly, the `unicorn/prefer-string-replace-all` rule has also been
disabled because it is not compatible with Node.js v14.18.0, which is
still supported by this package.