Adopt `eslint-plugin-unicorn` (#1812)
Start using `eslint-plugin-unicorn` to enforce opinionated stylistic
consistency. All rules are explicitly configured to ensure the code
follows the standard of this project. Also expand the coverage of
this and other general-purpose plugins from the `src/` code to all
JavaScript code.

This commit includes changes in compliance with these lint
changes.

Most rules are enabled as is. Some rules are disabled for various
reasons:
- `consistent-destructuring`: too situational to lint for in my opinion.
  Open to reconsider later.
- `import-style`: limited value while hard to configure. Default config
  not desired.
- `no-nested-ternary`: conflicts with Prettier.
- `no-null`: seems fine though I'm intrigued by the explanation given
  by the plugin. Open to reconsider later.
- `number-literal-case`: conflicts with Prettier.
- `prefer-spread`: imperfect rule that conflicts with this code base.
  Open to reconsider later.
- `prefer-string-raw`: I'm okay with escaping and am concerned about
  performance impacts of using `String.raw`.
- `prefer-ternary`: too situational to lint for in my opinion. Open to
  reconsider later.
- `prevent-abbreviations`: mainly don't have a solution for the "args"
  abbreviation. Open to reconsider later.

Lastly, the `unicorn/prefer-string-replace-all` rule is not in use
because it is not compatible with Node.js v14.18.0, which is still
supported by this package.