Configure ESLint's own rules
Update the ESLint config to enable and configure most of its own rules.
This iteration on the configuration is mostly focussed on enabling rules
the code base already complies with or can trivially comply with to
ensure consistency going forward. Any rule that has been disabled may be
reconsidered in the future.

Note that rules marked as "deprecated" by ESLint were not included here.