Configure ESLint's own rules (#1815)
Update the ESLint config to enable and configure most of its own rules.
This iteration on the configuration is mostly focussed on enabling rules
the code base already complies with or can trivially comply with to
ensure consistency going forward. Any rule that has been disabled may be
reconsidered in the future. Note that rules marked as "deprecated" by
ESLint were not included here.

Notably, the `prefer-object-has-own` rule was disabled because it is not
compatible with Node v14.18.0, which is still supported by this package.

This also re-enables the rule `ava/prefer-t-regex` disabled earlier because
of incompatibility in 56004b9b8b4f5ebc86a7d010cbc528877efb70ce.