Remove the dependency on knip (#1823)
Remove knip as it adds little value to this project as a continuous
check, instead it might still be used occasionally to review the
dependencies of this project. Along with the removal of the dependency
its configuration file is also removed. This removes the only jsonc file
from the project, which is reflected in other checks too.