Switch to Semgrep OSS (#1908)
Change how we run Semgrep in the CI from using `semgrep ci` to just
`semgrep`. This allows us to omit the Semgrep Application secret,
enabling us to run it on all changes (not just on `main`) at the cost of
a reduction in rules and no support for known vulnerability scanning (as
well as web application integration, which I don't really use anyway).
The former is a trade-off with the benefits this brings, the latter is
not a problem as this is covered elsewhere in this project.

So, the overall trade-off here is a reduction in security rules in
exchange for 1) running it on Pull Requests (because no secret is
required), 2) consistent results for all contributors (since the free
version is now used). To me, this is a clear win.