Correct escaping of `%` escaping for CMD (#1916)
Fix escaping of `%` for CMD (covering both the escape and qoute APIs) and at the same time simplify the logic. This removes the bit-flipping based on double quotes (`"`) and instead just escapes the double quotes. This handles the bug that was found by fuzzing (i.e. `"%_%`, which was generalized to `"%PATH%` to apply outside the CI environment too) and also _seems_ to handle all uses of `"` correctly (i.e., passing them to the program verbatim).