Add note about DEP0190 to the documentation
This has not been addressed in the invocations of the related APIs in
the tests because we still want to test these as long as we support a
Node.js version that supports calling the APIs in this way.
It has also not been addressed in the use of `spawnSync` in the scripts
of this project, which should be done as part of the bump to Node.js v24
for development.