Use `eslint-plugin-import`, don't use `dotenvx`
Following the update of `eslint-plugin-depend` this adjusts its
configuration to ignore the (new) alert for `dotenv` because the
suggested alternative does not look better for the purposes of this
project (and while I'd like to use Node.js' native solution, it's not
supported by all supported Node.js versions).

Secondly, this switches to `eslint-plugin-import` from its fork. This
is mainly a stability and popularity considerations - I'm open to
switching back to the fork if the original's reputation degrades.