Improve quoting for CMD (#1993)
This updates the escaping and quoting logic for CMD to make it more consistent.
The previous approach would break in some cases (e.g. with Maven/Java). The new
approach is perhaps more expected, actually quoting the entire argument in
double quotes and accounting for special character within those parameters.