Update fuzz logic to create corpus entry for new crashes
Implement logic to write any input for which fuzzing fails to a file in
the corpus automatically, simplifying the fuzzing process for developers
and update related logic accordingly.

Also do some refactoring in the fuzz logic to improve and simplify the
code and tests. Finally, get rid of the unused .corpus/ directory.