Update instructions and language related to fuzzing (#2015)
Update the instructions and language around fuzzing from that originally
written for jsfuzz by updated language and logic for fuzzing powered by
fast-check. In particular this updates the reusable fuzz workflow, contributing
guidelines, and some scripts and configuration files. The corpus has been
updated such that file names are the sha256 hash of their content, aligning it
with the newly introduced logic for adding new entries to the corpus when one
of the fuzz tests fails.

Also sneak in a couple of manual formatting fixes.