Add escaping support for BusyBox on Unix systems
The escaping is based off of Bash but was iteratively constructed using
fuzzing (still this implies it might be escaping unnecessarily in some
cases). Beyond fuzz testing (for >15k iterations), this addition is
properly unit, integration, and e2e tested (more on e2e below). The
breakage test explicitly exclude BusyBox because it was not yet
supported in the previous version (hence it behaves differently).

The e2e test setup is a bit complex because of the way BusyBox works. In
particular, invoking the busybox binary does not give you a shell (since
its a binary providing many functionalities) instead one must invoke
busybox as a file named "sh" to get its shell. Therefor, the end-to-end
tests create a (temporary) symbolic link for the project on-demand for
the end-to-end tests to use.

Lastly, the fuzz corpus is updated with a novel input.