Add escaping support for BusyBox on Unix systems (#2018)
The escaping is based off of Bash but was iteratively constructed using
fuzzing (still this implies it might be escaping unnecessarily in some
cases). Beyond fuzz testing, this addition is properly unit, integration, and
e2e tested. The breakage test explicitly exclude BusyBox because it was not yet
supported in the previous version (hence it behaves differently). The fuzz
corpus is even updated with a novel input.

One of the main challenges in implementing this was testing against the real
BusyBox binary. First, it needs to be invoked under the name `sh` to be able to
be used as a shell, so the end-to-end tests have been updated to account for
this (by adding a script that is ran before testing to create a sh->busybox
symlink). Second, it is the first Unix shell not (easily?) available on macOS,
so additional logic had to be implemented for that too.

Besides, this does some housekeeping to various parts of the project.