Integrate `dirty-waters` using the `dirty-waters-action` (#1883)
Add an integration for `dirty-waters` to continuously scan for, detect,
and address smells in the software supply chain. This provides a
systematic way of finding dependencies that lack certain transparency
measures such as an available source URL. This transparency can help
mitigate as well as investigate-after-the-fact attacks.

Certain smells were disabled:
- Deprecations: covered by `depreman`
- Forks: not interesting imho. Trivial for attackers to bypass, can be
  legitimate and might even be preferable to know about explicitly.
- Provenance: cool but adoption across npm is currently too low to
  enforce.
- Aliased packages: not a problem imho. To the extend it is a problem
  it can and should be addressed by tooling, not package authors.

All entries in the ignore section of the config file were rigorously
reviewed and are intentionally scoped narrowly. The only exception
is the `@types/.+` ignore which is there because this project does
not use tags (because its a giant monorepo) which I don't worry too
much about in general but especially for this project.