Disable npm cache when publishing (#2032)
Update the publishing workflow to avoid using a cache when publishing the
package to NPM. This is in order to avoid the published package from being
affected by a poisoned cache. This is meant is a hardening strategy and is
desired even if this attack _should_ be impossible with GitHub Actions (or its
current configuration).

Per the `actions/setup-cache` docs: "The cache input is optional, and
caching is turned off by default."