Try out `ghasum` (#2045)
Initialize ghasum by creating the `gha.sum` file with `ghasum init`.
Integrate it into the project by creating a local action to run
`ghasum verify` and use it in most jobs (some are unsupported) to
verify the checksums of all actions used.

For actions run after ghasum the tag has been changed from a commit
SHA to a version tag because the latter is more readable and the
benefits of the former are covered by ghasum. Unfortunately, some
actions are run before `ghasum` so they still need to use commit
SHAs as a means of guaranteeing integrity.