Set up `npm-package-json-lint`
Add `npm-package-json-lint` as a development dependency and use it to
lint the manifest, in particular enforce style and some preference rules
It adds to 3 other tools used for linting the manifest but does provide
some unique features (e.g. one I like is enforcing a certain range style
for different dependency categories).