Set up `npm-package-json-lint` (#2074)
Add `npm-package-json-lint` as a development dependency and use it to
lint the manifest, in particular enforce style and some preference rules
It adds to 3 other tools used for linting the manifest but does provide
some unique features (e.g. one I like is enforcing a certain range style
for different dependency categories).

The rules are mostly according to my preferences, partially guided by
the current practices of this project. Any rule that is currently disable
is open for discussion to be enabled - feel free to submit a change
request if you feel more rules should be enabled!

The configuration is formatted with spacing following the categories
present in the tool documentation.