Add support for which v4 and v5
Update the dependencies list to declare support for which v4 and v5,
which have the same API but compatibility with fewer Node.js versions.
Since this project still supports older Node.js versions, the multiple
version range support should enable us to make this a non-breaking
change.

The latest version is actually depended upon (see package-lock.json)
while compatibility with the older versions of which is continuously
tested in CI.