Depend on which v3.0.0, test against v4.0.0 and v5.0.0
Change how the support for which v3, v4, and v5 is maintained. Instead
of depending on the latest we depend on the earliest (through the
lockfile) and we test against newer versions. I believe this makes more
sense primarily because the (Node.js) compatibility testing uses a
version of which that's actually supposed to be compatible with all
tested Node.js versions.

Additionally, it prevents ls-engines from complaining.