Add support for `which` v4 and v5 (#2076)
* Add support for which v4 and v5

Update the dependencies list to declare support for which v4 and v5,
which have the same API but compatibility with fewer Node.js versions.
Since this project still supports older Node.js versions, the multiple
version range support should enable us to make this a non-breaking
change.

The latest version is actually depended upon (see package-lock.json)
while compatibility with the older versions of which is continuously
tested in CI.

* Depend on which v3.0.0, test against v4.0.0 and v5.0.0

Change how the support for which v3, v4, and v5 is maintained. Instead
of depending on the latest we depend on the earliest (through the
lockfile) and we test against newer versions. I believe this makes more
sense primarily because the (Node.js) compatibility testing uses a
version of which that's actually supposed to be compatible with all
tested Node.js versions.

Additionally, it prevents ls-engines from complaining.

* Update `check-runtime-deps.js`

require that the actually installed version is the earliest supported
version.

* Update the changelog