Update dirty-waters configuration
First, this disables the code signature check because it appears to be a
bit broken (`npm audit signature` reports all "1263 packages have
verified registry signatures" whereas dirty-waters says 35 don't, I
manually checked one of the packages it reports and it appears to have a
valid registry signature...) moreover it's not that meaningful as it's a
registry signature, not a signature from the package maintainer.

Second, this updates the configuration file with new ignores based on
the updated report. I briefly reviewed these reports (indicating it is
perhaps not the most useful signal...).