Create a "Threat Model"
Define a thread model and add it to the security policy. The threat
model sets out what is considered trusted and what isn't to help
security researchers, as well as users frankly, understand how this
library is intended to be used. It also allows for scrutiny of the
thus-far implicit thread model.