Mark `Shescape` export of `shescape/testing` as deprecated (#2103)
* Mark `Shescape` export of `shescape/testing` as deprecated

As `Stubscape` should be used instead. This alternative follows the
naming convention of the other stubs (e.g. `Failscape`) and makes it
explicit that it's not the actual `Shescape` class.

* Require that `@deprecated` tags have a description

Update the ESlint configuration to enforce that all `@deprecated` tags
have a description using the `match-description` rule from
eslint-plugin-jsdoc. Also relax the main description regular expression
to permit main descriptions that are not followed by any tags.