Prototype implementation of using linear regular expressions
... when available

Using a dedicated package I created for transparently creating regular
expressions with the `l` flag (when available). This requires some
refactoring to avoid significant performance overhead by pre-
initializing the regular expressions in a closure which can than reuse
it for repeated escaping.

For this proof of concept I just implemented it for one of the escape
modules. Also, it is not tested yet as I did not figure out a way to run
AVA tests with the `--enable-experimental-regexp-engine` flag.