Set up workflow to automatically update `gha.sum`
Add a new GitHub Actions workflow that will run on Dependabot updates
for GitHub Actions and automatically update the `gha.sum` file and push
the changes back.

This workflow exclusively uses commit SHAs because it does not verify
checksums. It is not possible to check ghasum checksums here because
they might be outdated when it runs and the goal is to update them,
hence the update would fail.

This updates the local `ghasum` action to NOT remove the `ghasum` binary
and instead add it to the path. This allows the new workflow to run the
`ghasum` command. Additionally, it adds the option to disable
verification (for the same reason).