Create a facade for `node:fs`
Localize the import surface with `node:fs` with an explicit facade.
Currently the facade API is 1-1 the (used part of) original API with the
exception of some simplifications.

In the future, this facade can help simplify migration as the original
API changes or (perhaps more likely in the case of a built-in library)
aid in writing code that supports both older and newer version of the
API when multiple runtime versions are supported.

One challenge, as is, is testing the facade (without actually
interacting with the underlying file system, unnecessarily). Hence,
there are currently no tests.