Ignore GHSA-5j98-mcp5-4vw2 (#2257)
A vulnerability affecting the CLI of glob. It's only used in development
dependencies, most likely only as a library.

The only dependency holding the update back is
`@stryker-mutator/tap-runner` (it is already upgraded upstream but given
there is no impact I respect them not releasing it).

The rest are some corrections for the changes introduced in 7b85656.