Enforce tree-like module hierarchy (for `src`)
Use `eslint-plugin-import` to enforce a tree-like module hierarchy on
the source code (i.e. the contents of `src`, other parts of the codebase
are excluded for now but may be refactored accordingly in the future).
The aim is to simplify the codebase now and going forward. Tree-like
hierarchies are easier to understand than arbitrary/graph-based module
structures and it also enforces a certain amount of encapsulation.