Create compatibility test suite for experimental regexp engine
Expand the `test/compat` suite with a dedicated test suite for ensuring
the regular expressions used in this project are compatible with the
experimental (linear-time) regexp engine. This is achieved by executing
(relevant) code with the `--enable-experimental-regexp-engine` CLI
option enabled and ensuring there are no unexpected runtime errors (as a
regular expression incompatible with the engine causes an error to be
thrown, as this commit demonstrates). In CI, this suite is run for all
supported version of Node.js because the exact set of supported regular
expression constructs by the experimental engine changes over time (and
we need to support all variants). The coverage report for this suite is
limited to only the relevant parts of the code base. The test suite
itself ensures it is executed with the required CLI options. This suite
has a custom runner -- like the pre-existing (runtime) compatibility
test suite -- for similar reasons (see header comment).