Partial support for using the linear regular expression engine (#2129)
Add support for using this package with the --enable-experimental-regexp-engine
option. It uses a dedicated package I created for transparently creating regular
expressions with the l flag when available. This requires some refactoring to
avoid performance overhead as a result of using the RegExp constructor,
initializing once and capturing the expression in a closure instead.

As of this commit the support is partial, covering only the no-shell use of
Shescape. Some of the affected regular expressions have been modified because
they were incompatible with the experimental regexp engine.

Compatibility with the experiment regular expression engine is tested using the
`npm run test:compat:regexp` command. The test suite can be found under
`test/compat/regexp-engine` and aims to execute all relevant uses of `RegExp`.
This test suite is executed on all supported Node.js engines as the set of
supported regular expression constructs changes over time. However, Node.js v14
is omitted from this because it did not yet have support for the experimental
regexp engine.

To accommodate the new `test:compat:regexp` script, related scripts have been
refactored. The original `test:compat` script has been renamed to
`test:compat:runtime` and the `test:compat` script now executes all
compatibility tests.