Attempt fixing vulnerable dependencies when detected
Extend the `audit-dev.yml` job "Vulnerabilities" to try and fix the
vulnerabilities with `npm audit fix` and submit a PR for the changes if
any vulnerabilities are detected on a schedule (we don't want this for
non-schedule jobs because it would get messy, e.g. creating PRs to merge
the fix into (unrelated) development branches).